tasks_api/pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["ksredkin <ksredkin0968@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = "^0.124.4"
uvicorn = "^0.38.0"
python-jose = "^3.5.0"
dotenv = "^0.9.9"
httpx = "^0.28.1"
bcrypt = "4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.11"
sqlalchemy = "^2.0.46"
alembic = "^1.18.3"
aiogram = "^3.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
